New parsing logic.
Parse CSV into a locally running instance of Postgres, into the same
database used by a locally running instance of the Ohana API server.

This skips the export step because the Ohana server can export the data
to a dumpfile itself.

The db connection details will have to be customized for your local
system.